Update detector module name
