--- art/defences/detection/evasion/__init__.py
@@ -1,4 +1,0 @@
-"""
-Module providing methods for detecting adversarial samples under a common interface.
-"""
-from art.defences.detection.evasion.detector import BinaryInputDetector, BinaryActivationDetector
