+++ art/defences/detector/evasion/__init__.py
@@ -1,0 +1,4 @@
+"""
+Module providing methods for detecting adversarial samples under a common interface.
+"""
+from art.defences.detector.evasion.detector import BinaryInputDetector, BinaryActivationDetector
