+++ art/defences/detector/evasion/subsetscanning/detector.py
@@ -30,7 +30,7 @@ import six
 from art.estimators.estimator import BaseEstimator, NeuralNetworkMixin, LossGradientsMixin
 from art.estimators.classification.classifier import ClassifierMixin, ClassGradientsMixin
 
-from art.defences.detection.evasion import Scanner
+from art.defences.detector.evasion import Scanner
 
 logger = logging.getLogger(__name__)
