+++ art/defences/detector/poison/__init__.py
@@ -2,4 +2,4 @@
 Poison detection defence API. Use the :class:`.PoisonFilteringDefence` wrapper to be able to apply a defence for a
 preexisting model.
 """
-from art.defences.detection.poison.ground_truth_evaluator import GroundTruthEvaluator
+from art.defences.detector.poison.ground_truth_evaluator import GroundTruthEvaluator
