+++ art/defences/detector/poison/ground_truth_evaluator.py
