+++ tests/defences/detector/poison/test_clustering_analyzer.py
@@ -22,7 +22,7 @@ import unittest
 
 import numpy as np
 
-from art.defences.detection.poison import ClusteringAnalyzer
+from art.defences.detector.poison import ClusteringAnalyzer
 
 logger = logging.getLogger(__name__)
