+++ tests/defences/detector/poison/test_ground_truth_evaluator.py
@@ -22,7 +22,7 @@ import logging
 import pprint
 import unittest
 
-from art.defences.detection.poison import GroundTruthEvaluator
+from art.defences.detector.poison import GroundTruthEvaluator
 
 from tests.utils import master_seed
