+++ tests/defences/detector/poison/test_provenance_defence.py
@@ -25,25 +25,29 @@ from sklearn.svm import SVC
 
 from art.attacks.poisoning.poisoning_attack_svm import PoisoningAttackSVM
 from art.estimators.classification.scikitlearn import SklearnClassifier, ScikitlearnSVC
-from art.defences.detection.poison.roni import RONIDefense
+from art.defences.detector.poison.provenance_defense import ProvenanceDefense
 from art.utils import load_mnist
 
 from tests.utils import master_seed
 
 logger = logging.getLogger(__name__)
 
-NB_TRAIN, NB_POISON, NB_VALID, NB_TRUSTED = 40, 5, 40, 15
+NB_TRAIN = 40
+NB_POISON = 5
+NB_VALID = 10
+NB_TRUSTED = 10
+NB_DEVICES = 4
 kernel = "linear"
 
 
-class TestRONI(unittest.TestCase):
+class TestProvenanceDefence(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        master_seed(seed=1234)
+        master_seed(seed=301)
         (x_train, y_train), (x_test, y_test), min_, max_ = load_mnist()
         y_train = np.argmax(y_train, axis=1)
         y_test = np.argmax(y_test, axis=1)
-        zero_or_four = np.logical_or(y_train == 4, y_train == 0)
+        zero_or_four = np.logical_or(y_train == 4, y_train == 0, y_train == 9)
         x_train = x_train[zero_or_four]
         y_train = y_train[zero_or_four]
         tr_labels = np.zeros((y_train.shape[0], 2))
@@ -78,6 +82,9 @@ class TestRONI(unittest.TestCase):
         x_test = x_test[NB_VALID:]
         y_test = y_test[NB_VALID:]
 
+        clean_prov = np.random.randint(NB_DEVICES - 1, size=x_train.shape[0])
+        p_train = np.eye(NB_DEVICES)[clean_prov]
+
         no_defense = ScikitlearnSVC(model=SVC(kernel=kernel, gamma="auto"), clip_values=(min_, max_))
         no_defense.fit(x=x_train, y=y_train)
         poison_points = np.random.randint(no_defense._model.support_vectors_.shape[0], size=NB_POISON)
@@ -100,10 +107,13 @@ class TestRONI(unittest.TestCase):
         # Stack on poison to data and add provenance of bad actor
         all_data = np.vstack([x_train, poisoned_data])
         all_labels = np.vstack([y_train, poison_labels])
+        poison_prov = np.zeros((NB_POISON, NB_DEVICES))
+        poison_prov[:, NB_DEVICES - 1] = 1
+        all_p = np.vstack([p_train, poison_prov])
 
         model = SVC(kernel=kernel, gamma="auto")
         cls.mnist = (
-            (all_data, all_labels),
+            (all_data, all_labels, all_p),
             (x_test, y_test),
             (trusted_data, trusted_labels),
             (valid_data, valid_labels),
@@ -112,45 +122,54 @@ class TestRONI(unittest.TestCase):
         cls.classifier = SklearnClassifier(model=model, clip_values=(min_, max_))
 
         cls.classifier.fit(all_data, all_labels)
-        cls.defense_cal = RONIDefense(
-            cls.classifier, all_data, all_labels, trusted_data, trusted_labels, eps=0.1, calibrated=True
+        cls.defence_trust = ProvenanceDefense(
+            cls.classifier, all_data, all_labels, all_p, x_val=trusted_data, y_val=trusted_labels, eps=0.1
         )
-        cls.defence_no_cal = RONIDefense(
-            cls.classifier, all_data, all_labels, trusted_data, trusted_labels, eps=0.1, calibrated=False
-        )
+        cls.defence_no_trust = ProvenanceDefense(cls.classifier, all_data, all_labels, all_p, eps=0.1)
 
     def setUp(self):
-        master_seed(seed=1234)
+        master_seed(seed=301)
 
     def test_wrong_parameters_1(self):
-        self.assertRaises(ValueError, self.defence_no_cal.set_params, eps=-2.0)
-        self.assertRaises(ValueError, self.defense_cal.set_params, eps=-2.0)
+        self.assertRaises(ValueError, self.defence_no_trust.set_params, eps=-2.0)
+        self.assertRaises(ValueError, self.defence_trust.set_params, eps=-2.0)
 
     def test_wrong_parameters_2(self):
-        (all_data, _), (_, y_test), (_, _), (_, _), (_, _) = self.mnist
-        self.assertRaises(ValueError, self.defence_no_cal.set_params, x_train=-all_data, y_train=y_test)
-        self.assertRaises(ValueError, self.defense_cal.set_params, x_train=-all_data, y_train=y_test)
+        self.assertRaises(ValueError, self.defence_no_trust.set_params, pp_valid=-0.1)
+        self.assertRaises(ValueError, self.defence_trust.set_params, eps=-0.1)
+
+    def test_wrong_parameters_3(self):
+        (all_data, _, _), (_, y_test), (_, _), (_, _), (_, _) = self.mnist
+        self.assertRaises(ValueError, self.defence_no_trust.set_params, x_train=-all_data, y_train=y_test)
+        self.assertRaises(ValueError, self.defence_trust.set_params, x_train=-all_data, y_train=y_test)
+
+    def test_wrong_parameters_4(self):
+        (_, _, p_train), (x_test, y_test), (_, _), (_, _), (_, _) = self.mnist
+        self.assertRaises(
+            ValueError, self.defence_no_trust.set_params, x_train=-x_test, y_train=y_test, p_train=p_train
+        )
+        self.assertRaises(ValueError, self.defence_trust.set_params, x_train=-x_test, y_train=y_test, p_train=p_train)
 
     def test_detect_poison(self):
-        _, clean_trust = self.defense_cal.detect_poison()
-        _, clean_no_trust = self.defence_no_cal.detect_poison()
+        _, clean_trust = self.defence_trust.detect_poison()
+        _, clean_no_trust = self.defence_no_trust.detect_poison()
         real_clean = np.array([1 if i < NB_TRAIN else 0 for i in range(NB_TRAIN + NB_POISON)])
-        pc_tp_cal = np.average(real_clean[:NB_TRAIN] == clean_trust[:NB_TRAIN])
-        pc_tn_cal = np.average(real_clean[NB_TRAIN:] == clean_trust[NB_TRAIN:])
-        self.assertGreaterEqual(pc_tn_cal, 0)
-        self.assertGreaterEqual(pc_tp_cal, 0.7)
+        pc_tp_trust = np.average(real_clean[:NB_TRAIN] == clean_trust[:NB_TRAIN])
+        pc_tn_trust = np.average(real_clean[NB_TRAIN:] == clean_trust[NB_TRAIN:])
+        self.assertGreaterEqual(pc_tn_trust, 0.7)
+        self.assertGreaterEqual(pc_tp_trust, 0.7)
 
-        pc_tp_no_cal = np.average(real_clean[:NB_TRAIN] == clean_no_trust[:NB_TRAIN])
-        pc_tn_no_cal = np.average(real_clean[NB_TRAIN:] == clean_no_trust[NB_TRAIN:])
-        self.assertGreaterEqual(pc_tn_no_cal, 0)
-        self.assertGreaterEqual(pc_tp_no_cal, 0.7)
+        pc_tp_no_trust = np.average(real_clean[:NB_TRAIN] == clean_no_trust[:NB_TRAIN])
+        pc_tn_no_trust = np.average(real_clean[NB_TRAIN:] == clean_no_trust[NB_TRAIN:])
+        self.assertGreaterEqual(pc_tn_no_trust, 0.7)
+        self.assertGreaterEqual(pc_tp_no_trust, 0.7)
 
     def test_evaluate_defense(self):
         real_clean = np.array([1 if i < NB_TRAIN else 0 for i in range(NB_TRAIN + NB_POISON)])
-        self.defence_no_cal.detect_poison()
-        self.defense_cal.detect_poison()
-        logger.info(self.defense_cal.evaluate_defence(real_clean))
-        logger.info(self.defence_no_cal.evaluate_defence(real_clean))
+        self.defence_no_trust.detect_poison()
+        self.defence_trust.detect_poison()
+        logger.info(self.defence_trust.evaluate_defence(real_clean))
+        logger.info(self.defence_no_trust.evaluate_defence(real_clean))
 
 
 if __name__ == "__main__":
